added comment remover for java and imported stylyng tool for java files

+++ src/ccaligner.py
@@ -1,7 +1,9 @@
 import lexical_analysis.pretty_printing as prpr
 from launching import define_args
 from shutil import rmtree
+import os
 from clone_detection.algorithm import CCalignerAlgorithm
+
 
 args = define_args()
 
@@ -14,13 +16,18 @@ elif language == 'java':
 
 
 pretty_loc = "./data/normalized_codebases/"
-rmtree(pretty_loc)
+if os.path.exists(pretty_loc):
+    rmtree(pretty_loc)
 
+if (language == 'python'):
+    pp = prpr.PrettyPrinterPy(codebase_loc, pretty_loc, language)
+elif (language == 'java'):
+    pp = prpr.PrettyPrinterJava(codebase_loc, pretty_loc, language)
 
-pp = prpr.PrettyPrinter(codebase_loc, pretty_loc, language)
 pp.pretty_print()
 
-final_dir = pretty_loc + codebase_loc.split('/')[-1] + '/' + 'obfuscated'
+
+"""final_dir = pretty_loc + codebase_loc.split('/')[-1] + '/' + 'obfuscated'
 
 cca = CCalignerAlgorithm(final_dir, lang_ext, 3, 0)
 pairs = cca.run_algo()
@@ -29,4 +36,4 @@ for file1, file2 in pairs:
     file_name2 = file2.split('/')[-2]
     fragment1 = file1.split('/')[-1][:-3]
     fragment2 = file2.split('/')[-1][:-3]
-    print(f"{file_name1} and {file_name2} contain codeclone in lines {fragment1} and {fragment2} respectively")
+    print(f"{file_name1} and {file_name2} contain codeclone in lines {fragment1} and {fragment2} respectively")"""
